Tasks/testing

+++ controllers/tasks/removeById.js
@@ -2,6 +2,15 @@ const Task = require("../../models/tasks");
 
 const removeById = async (req, res) => {
   const { taskId } = req.params;
+  const { id: owner } = req.user;
+
+  const isOwner = await Task.findOne({ _id: taskId, owner });
+  
+
+
+  if (!isOwner) {
+    return res.status(403).json({ message: "This is not your task" });
+  }
   const result = await Task.findByIdAndRemove(taskId);
 
   if (!result) {
